+++ processor/transformprocessor/internal/common/func_replace_all_patterns_test.go
@@ -27,7 +27,7 @@ func Test_replaceAllPatterns(t *testing.T) {
 	input := pcommon.NewMap()
 	input.InsertString("test", "hello world")
 	input.InsertString("test2", "hello")
-	input.InsertString("test3", "goodbye")
+	input.InsertString("test3", "goodbye world1 and world2")
 
 	target := &testGetSetter{
 		getter: func(ctx TransformContext) interface{} {
@@ -55,7 +55,7 @@ func Test_replaceAllPatterns(t *testing.T) {
 				expectedMap.Clear()
 				expectedMap.InsertString("test", "hello {universe} world")
 				expectedMap.InsertString("test2", "hello {universe}")
-				expectedMap.InsertString("test3", "goodbye")
+				expectedMap.InsertString("test3", "goodbye world1 and world2")
 			},
 		},
 		{
@@ -67,7 +67,19 @@ func Test_replaceAllPatterns(t *testing.T) {
 				expectedMap.Clear()
 				expectedMap.InsertString("test", "hello world")
 				expectedMap.InsertString("test2", "hello")
-				expectedMap.InsertString("test3", "goodbye")
+				expectedMap.InsertString("test3", "goodbye world1 and world2")
+			},
+		},
+		{
+			name:        "multiple regex match",
+			target:      target,
+			pattern:     `world[^\s]*(\s?)`,
+			replacement: "**** ",
+			want: func(expectedMap pcommon.Map) {
+				expectedMap.Clear()
+				expectedMap.InsertString("test", "hello **** ")
+				expectedMap.InsertString("test2", "hello")
+				expectedMap.InsertString("test3", "goodbye **** and **** ")
 			},
 		},
 	}
@@ -131,4 +143,21 @@ func Test_replaceAllPatterns_get_nil(t *testing.T) {
 	exprFunc, err := replaceAllPatterns(target, "regexp", "{anything}")
 	assert.Nil(t, err)
 	exprFunc(ctx)
+}
+
+func Test_replaceAllPatterns_invalid_pattern(t *testing.T) {
+	target := &testGetSetter{
+		getter: func(ctx TransformContext) interface{} {
+			t.Errorf("nothing should be received in this scenario")
+			return nil
+		},
+		setter: func(ctx TransformContext, val interface{}) {
+			t.Errorf("nothing should be set in this scenario")
+		},
+	}
+
+	invalidRegexPattern := "*"
+	exprFunc, err := replaceAllPatterns(target, invalidRegexPattern, "{anything}")
+	assert.Nil(t, exprFunc)
+	assert.Contains(t, err.Error(), "error parsing regexp:")
 }